contact page setup, needs more content, got the theme workinggit add ., content changes and code tidy up
