+++ src/pages/contact/contact-styles.js
@@ -1,0 +1,14 @@
+// import { makeStyles, Theme } from "@material-ui/core";
+
+
+// export const useStyles = makeStyles((theme: Theme) => ({
+//     root: {
+//         maxWidth: 345,
+//         maxHeight: 345,
+//     },
+//     cardWidth: {
+//         maxWidth: '35rem',
+//         color: theme.cream,
+//     },
+// }) { index: 1 },
+// );
